+++ component/component_test.js
@@ -1564,6 +1564,59 @@ steal("can/util/vdom/document", "can/util/vdom/build_fragment","can", "can/map/d
 
 			can.append(this.$fixture, template());
 		});
+
+		test("%root property is available in a viewModel", function () {
+			var viewModel = can.Map.extend({});
+
+			can.Component.extend({
+				tag: "foo",
+				viewModel: viewModel,
+				init: function () {
+					ok(this.viewModel.attr('%root'), "viewModel contains '%root' property");
+				}
+			});
+
+			var template = can.stache("<foo/>");
+
+			can.append(this.$fixture, template());
+		});
+
+		test('Type in a component’s viewModel doesn’t work correctly with lists (#2250)', function() {
+			var Item = can.Map.extend({
+				define: {
+					prop: {
+						value: true
+					}
+				}
+			});
+
+			var Collection = can.List.extend({
+				Map: Item
+			}, {});
+
+			can.Component.extend({
+				tag: "test-component",
+				viewModel: {
+					define: {
+						collection: {
+							Type: Collection
+						},
+						vmProp: {
+							get: function() {
+								return this.attr('collection.0.prop');
+							}
+						}
+					}
+				}
+			});
+
+			var frag = can.stache('<test-component collection="{collection}"></test-component>')({
+				collection: [{}]
+			});
+			var vmPropVal = can.viewModel(can.$(frag.firstChild)).attr('vmProp');
+			ok(vmPropVal, 'value is from defined prop');
+		});
+
 		// PUT NEW TESTS ABOVE THIS LINE
 	}
 
@@ -1756,7 +1809,7 @@ steal("can/util/vdom/document", "can/util/vdom/build_fragment","can", "can/map/d
 
 			// Dispatches async
 			setTimeout(function() {
-				equal(count, 4, '2 items unbound, but unbinding the temporary unbind and the actual unbind');
+				equal(count, 2, '2 items unbound. Previously, this would unbind 4 times because of switching to fast path');
 				can.unbindAndTeardown = old;
 
 				start();
@@ -1797,8 +1850,15 @@ steal("can/util/vdom/document", "can/util/vdom/build_fragment","can", "can/map/d
 			map.attr("show", false);
 			map.attr("show", true);
 			map.attr("show", false);
-			equal(map._bindings,1, "only one binding");
-			can.remove(can.$("#qunit-fixture>*"));
+			map.attr("show", true);
+			map.attr("show", false);
+			stop();
+			setTimeout(function(){
+				equal(map._bindings,1, "only one binding");
+				can.remove(can.$("#qunit-fixture>*"));
+				start();
+			},10);
+			
 		});
 		
 		// PUT NEW TESTS THAT NEED TO TEST AGAINST STACHE JUST ABOVE THIS LINE