+++ view/stache/stache.js
@@ -234,11 +234,12 @@ steal(
 						if( !state.node.attributes ) {
 							state.node.attributes = [];
 						}
-						state.node.attributes.push(function(scope, options){
+						state.node.attributes.push(function(scope, options, nodeList){
 							attrCallback(this,{
 								attributeName: attrName,
 								scope: scope,
-								options: options
+								options: options,
+								nodeList: nodeList
 							});
 						});
 					}
@@ -267,7 +268,15 @@ steal(
 				
 				
 				if(expression === "else") {
-					(state.attr && state.attr.section ? state.attr.section : state.textContentOnly || section).inverse();
+					var inverseSection;
+					if(state.attr && state.attr.section) {
+						inverseSection = state.attr.section;
+					} else if(state.node && state.node.section ) {
+						inverseSection = state.node.section;
+					} else {
+						inverseSection = state.textContentOnly || section;
+					}
+					inverseSection.inverse();
 					return;
 				}