+++ changelog.md
@@ -3,6 +3,64 @@
 @parent guides 5
 -->
 
+__2.3.18__ ( Mar 03 2016 )
+
+- [Make getter / setter converter functions possible with stache helpers](https://github.com/canjs/canjs/issues/2299)
+- [Fix uppercase html parsing](https://github.com/canjs/canjs/pull/2296)
+- [Move where %root is added to a component’s initial viewModel data](https://github.com/canjs/canjs/pull/2282)
+- [`Type` in a component’s VM doesn’t work correctly with lists](https://github.com/canjs/canjs/issues/2250)
+
+
+__2.3.17__ ( Feb 19 2016 )
+
+- [Documentation for can.Deferred reject is not correct](https://github.com/canjs/canjs/issues/2280)
+- [Properties on a can.route.map are coerced to strings when they should not be](https://github.com/canjs/canjs/issues/2277)
+- [can-value does not teardown binding](https://github.com/canjs/canjs/issues/2270)
+- [Stache parser creating incorrect output](https://github.com/canjs/canjs/issues/2229)
+- [console.log.apply not supported on IE11?](https://github.com/canjs/canjs/issues/2227)
+
+
+__2.3.16__ ( Feb 16 2016 )
+
+- [Documentation for can.List.unshift.apply is not correct](https://github.com/canjs/canjs/issues/2262)
+- [Switch define docs to use new example version.](https://github.com/canjs/canjs/pull/2261)
+- [Batched change notifies child component before parent stache binding](https://github.com/canjs/canjs/issues/2252)
+- [../ looks up wrong property on maps](https://github.com/canjs/canjs/issues/2244)
+- [can.Map constructor is observable](https://github.com/canjs/canjs/issues/2220)
+- [Route/Map triggering too many changes](https://github.com/canjs/canjs/issues/2206)
+- [{{#if}} statement within partial gets displayed outside partial](https://github.com/canjs/canjs/issues/2186)
+
+
+__2.3.15__ ( Feb 15 2016 )
+
+- [adds jquery as a dep](https://github.com/canjs/canjs/pull/2242)
+
+
+__2.3.14__ ( Feb 05 2016 )
+
+- [double get](https://github.com/canjs/canjs/pull/2239)
+- [A trapped and &quot;not observed&quot; observed info object would still be observed.](https://github.com/canjs/canjs/pull/2236)
+- [2190 canjs api update](https://github.com/canjs/canjs/pull/2232)
+- [Define getters run twice. setVal is missing on first run.](https://github.com/canjs/canjs/issues/2230)
+- [Batching/Binding Issue](https://github.com/canjs/canjs/issues/2223)
+- [Fix `map.attr` with empty string.](https://github.com/canjs/canjs/pull/2219)
+- [Set jQuery package version using Caret ](https://github.com/canjs/canjs/issues/2212)
+- [2183 update guides](https://github.com/canjs/canjs/pull/2184)
+- [@index is double wrapped compute object in mustache/stache helper in latest version (V2.3.8)](https://github.com/canjs/canjs/issues/2179)
+
+
+__2.3.12/2.3.13__ ( Feb 01 2016 )
+
+- [jquery 2.2 in peer deps not supported](https://github.com/canjs/canjs/issues/2207)
+- [Keys aren&#39;t documented](https://github.com/canjs/canjs/issues/2202)
+- [calling methods directly on an instance of a can.Map with can.compute.read](https://github.com/canjs/canjs/issues/2199)
+- [Parser is not allowing anchor tags as child of inline elements](https://github.com/canjs/canjs/issues/2169)
+- [can.viewModel mentioned in the navigation but not the root page contents](https://github.com/canjs/canjs/issues/2165)
+- [Asynchronous virtual properties cause extra recomputes](https://github.com/canjs/canjs/issues/1915)
+- [Child bindings are called before the parent in Stache](https://github.com/canjs/canjs/issues/1866)
+- [Dynamic select menus don&#39;t live bind correctly](https://github.com/canjs/canjs/issues/1762)
+
+
 __2.3.11__ ( Jan 21 2016 )
 
 - [makes events fired as a result of a batch happen after any previous batches](https://github.com/canjs/canjs/pull/2203)